Move the treewalker printer to the treewalker module
